Address PR comments
Rename ForeignObjectLabel -> HtmlLabel

Split out HtmlLabel and LabelAnchorLine to separate components.

Export HtmlLabel as a separate component as it uses a small subset of the shared props.
