+++ packages/visx-annotation/src/index.ts
@@ -1,5 +1,6 @@
 export { default as Connector } from './components/Connector';
 export { default as Label } from './components/Label';
+export { default as HtmlLabel } from './components/HtmlLabel';
 export { default as CircleSubject } from './components/CircleSubject';
 export { default as LineSubject } from './components/LineSubject';
 export { default as Annotation } from './components/Annotation';
