+++ packages/visx-demo/src/pages/docs/annotation.tsx
@@ -6,6 +6,7 @@ import CircleSubject from '../../../../visx-annotation/src/components/CircleSubj
 import LineSubject from '../../../../visx-annotation/src/components/LineSubject';
 import Connector from '../../../../visx-annotation/src/components/Connector';
 import Label from '../../../../visx-annotation/src/components/Label';
+import HtmlLabel from '../../../../visx-annotation/src/components/HtmlLabel';
 import LinePathAnnotationDeprecated from '../../../../visx-annotation/src/deprecated/LinePathAnnotation';
 import DocPage from '../../components/DocPage';
 import AnnotationTile from '../../components/Gallery/AnnotationTile';
@@ -17,6 +18,7 @@ const components = [
   LineSubject,
   Connector,
   Label,
+  HtmlLabel,
   LinePathAnnotationDeprecated,
 ];
